vector-pipeline/requirements.txt
aiobotocore==2.24.2
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aioitertools==0.12.0
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.10.0
apache-beam==2.67.0
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
attrs==25.3.0
botocore==1.40.18
certifi==2025.8.3
cffi==2.0.0
charset-normalizer==3.4.3
click==8.3.0
colorama==0.4.6
crcmod==1.7
dill==0.3.1.1
dnspython==2.8.0
docopt==0.6.2
et_xmlfile==2.0.0
fastavro==1.12.0
fasteners==0.20
filelock==3.19.1
frozenlist==1.7.0
fsspec==2025.9.0
grpcio==1.65.5
h11==0.16.0
h2==4.3.0
hdfs==2.7.3
hpack==4.1.0
httpcore==1.0.9
httplib2==0.22.0
httpx==0.28.1
huggingface-hub==0.35.0
hyperframe==6.1.0
idna==3.10
Jinja2==3.1.6
jmespath==1.0.1
joblib==1.5.2
jsonpickle==3.4.2
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
MarkupSafe==3.0.2
minio==7.2.16
mpmath==1.3.0
msgpack==1.1.1
multidict==6.6.4
networkx==3.5
numpy==2.2.6
objsize==0.7.1
openpyxl==3.1.5
orjson==3.11.3
packaging==25.0
pandas==2.3.2
pillow==11.3.0
portalocker==3.2.0
propcache==0.3.2
proto-plus==1.26.1
protobuf==5.29.5
psycopg2-binary==2.9.10
pyarrow==18.1.0
pyarrow-hotfix==0.7
pycparser==2.23
pycryptodome==3.23.0
pydantic==2.11.9
pydantic_core==2.33.2
pydot==1.4.2
PyJWT==2.10.1
pymilvus==2.5.16
pymongo==4.15.1
pyparsing==3.2.4
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
pywin32==311
PyYAML==6.0.2
qdrant-client==1.15.1
ray==2.49.2
redis==5.3.1
referencing==0.36.2
regex==2025.9.18
requests==2.32.5
rpds-py==0.27.1
s3fs==2025.9.0
safetensors==0.6.2
scikit-learn==1.7.2
scipy==1.16.2
sentence-transformers==5.1.1
setuptools==80.9.0
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
sympy==1.14.0
threadpoolctl==3.6.0
tokenizers==0.22.1
torch==2.8.0
tqdm==4.67.1
transformers==4.56.2
typing-inspection==0.4.1
typing_extensions==4.15.0
tzdata==2025.2
ujson==5.11.0
urllib3==2.5.0
wrapt==1.17.3
yarl==1.20.1
zstandard==0.25.0
